Revert "[add] location box"
This reverts commit bca9440492f8048d2d54b56ecf2e33c99614ada3.

+++ backend/src/routes/location.rs
@@ -12,7 +12,7 @@ use crate::{
 
 #[derive(Serialize, Debug)]
 struct LocationResponse {
-    location: HashMap<String, i64>,
+    location: HashMap<String, i32>,
 }
 
 #[get("/api/location?<character_id>")]
